fix: address CodeRabbit feedback and consolidate duplicate endpoints
