+++ Cargo.toml
@@ -71,7 +71,7 @@ askama = "0.12"
 askama_axum = "0.4"
 
 # HTTP client and utilities
-reqwest = "0.11"
+reqwest = "0.12"
 hyper = "1.0"
 
 # Security and authentication
