+++ vibe-ensemble-web/src/handlers/api.rs
@@ -1,6 +1,6 @@
-//! API handlers
+//! Additional API handlers for knowledge and message management
 
-use crate::{websocket::WebSocketManager, Error, Result};
+use crate::{Error, Result};
 use axum::{
     extract::{Path, Query, State},
     http::StatusCode,
@@ -11,272 +11,15 @@ use serde_json::{json, Value};
 use std::{collections::HashMap, sync::Arc};
 use uuid::Uuid;
 use vibe_ensemble_core::{
-    agent::Agent,
-    issue::{Issue, IssuePriority, IssueStatus},
     knowledge::KnowledgeEntry,
     message::Message,
 };
 use vibe_ensemble_storage::StorageManager;
 
-/// Health check endpoint
-pub async fn health(State(storage): State<Arc<StorageManager>>) -> Result<Json<Value>> {
-    // Check database health
-    storage.health_check().await?;
-
-    Ok(Json(json!({
-        "status": "healthy",
-        "timestamp": chrono::Utc::now(),
-    })))
-}
-
-/// System statistics endpoint
-pub async fn stats(State(storage): State<Arc<StorageManager>>) -> Result<Json<Value>> {
-    let stats = storage.stats().await?;
-
-    Ok(Json(json!({
-        "agents": stats.agents_count,
-        "issues": stats.issues_count,
-        "messages": stats.messages_count,
-        "knowledge": stats.knowledge_count,
-        "prompts": stats.prompts_count,
-        "timestamp": chrono::Utc::now(),
-    })))
-}
-
 // ======================
-// AGENT API ENDPOINTS
+// KNOWLEDGE API ENDPOINTS  
 // ======================
 
-/// Query parameters for agent listing
-#[derive(Debug, Deserialize)]
-pub struct AgentQuery {
-    pub limit: Option<i32>,
-    pub offset: Option<i32>,
-    pub status: Option<String>,
-    pub agent_type: Option<String>,
-}
-
-/// List all agents with optional filtering
-pub async fn agents_list(
-    State(storage): State<Arc<StorageManager>>,
-    Query(query): Query<AgentQuery>,
-) -> Result<Json<Value>> {
-    let agents = storage.agents().list().await?;
-
-    // Apply basic filtering
-    let filtered_agents: Vec<Agent> = agents
-        .into_iter()
-        .filter(|agent| {
-            if let Some(status) = &query.status {
-                format!("{:?}", agent.status).to_lowercase() == status.to_lowercase()
-            } else {
-                true
-            }
-        })
-        .filter(|agent| {
-            if let Some(agent_type) = &query.agent_type {
-                format!("{:?}", agent.agent_type).to_lowercase() == agent_type.to_lowercase()
-            } else {
-                true
-            }
-        })
-        .skip(query.offset.unwrap_or(0) as usize)
-        .take(query.limit.unwrap_or(100) as usize)
-        .collect();
-
-    Ok(Json(json!({
-        "agents": filtered_agents,
-        "total": filtered_agents.len(),
-        "timestamp": chrono::Utc::now(),
-    })))
-}
-
-/// Get specific agent details
-pub async fn agent_detail(
-    State(storage): State<Arc<StorageManager>>,
-    Path(id): Path<Uuid>,
-) -> Result<Json<Value>> {
-    let agent = storage
-        .agents()
-        .find_by_id(id)
-        .await?
-        .ok_or_else(|| Error::NotFound(format!("Agent with id {}", id)))?;
-
-    Ok(Json(json!({
-        "agent": agent,
-        "timestamp": chrono::Utc::now(),
-    })))
-}
-
-// ======================
-// ISSUE API ENDPOINTS
-// ======================
-
-/// Issue creation/update request
-#[derive(Debug, Deserialize)]
-pub struct IssueRequest {
-    pub title: String,
-    pub description: String,
-    pub priority: String,
-    pub assigned_agent_id: Option<Uuid>,
-}
-
-/// Query parameters for issue listing
-#[derive(Debug, Deserialize)]
-pub struct IssueQuery {
-    pub limit: Option<i32>,
-    pub offset: Option<i32>,
-    pub status: Option<String>,
-    pub priority: Option<String>,
-    pub assigned_agent_id: Option<Uuid>,
-}
-
-/// List all issues with optional filtering
-pub async fn issues_list(
-    State(storage): State<Arc<StorageManager>>,
-    Query(query): Query<IssueQuery>,
-) -> Result<Json<Value>> {
-    let issues = storage.issues().list().await?;
-
-    // Apply basic filtering
-    let filtered_issues: Vec<Issue> = issues
-        .into_iter()
-        .filter(|issue| {
-            if let Some(status) = &query.status {
-                format!("{:?}", issue.status).to_lowercase() == status.to_lowercase()
-            } else {
-                true
-            }
-        })
-        .filter(|issue| {
-            if let Some(priority) = &query.priority {
-                format!("{:?}", issue.priority).to_lowercase() == priority.to_lowercase()
-            } else {
-                true
-            }
-        })
-        .filter(|issue| {
-            if let Some(agent_id) = query.assigned_agent_id {
-                issue.assigned_agent_id == Some(agent_id)
-            } else {
-                true
-            }
-        })
-        .skip(query.offset.unwrap_or(0) as usize)
-        .take(query.limit.unwrap_or(100) as usize)
-        .collect();
-
-    Ok(Json(json!({
-        "issues": filtered_issues,
-        "total": filtered_issues.len(),
-        "timestamp": chrono::Utc::now(),
-    })))
-}
-
-/// Create a new issue
-pub async fn issue_create(
-    State(storage): State<Arc<StorageManager>>,
-    State(ws_manager): State<Arc<WebSocketManager>>,
-    Json(request): Json<IssueRequest>,
-) -> Result<(StatusCode, Json<Value>)> {
-    let priority = match request.priority.as_str() {
-        "Low" => IssuePriority::Low,
-        "Medium" => IssuePriority::Medium,
-        "High" => IssuePriority::High,
-        "Critical" => IssuePriority::Critical,
-        _ => return Err(Error::BadRequest("Invalid priority".to_string())),
-    };
-
-    let mut issue = Issue::new(request.title.clone(), request.description, priority)?;
-    issue.assigned_agent_id = request.assigned_agent_id;
-
-    storage.issues().create(&issue).await?;
-
-    // Broadcast to WebSocket clients
-    let _ = ws_manager.broadcast_issue_created(issue.id, request.title, format!("{:?}", priority));
-
-    Ok((
-        StatusCode::CREATED,
-        Json(json!({
-            "issue": issue,
-            "timestamp": chrono::Utc::now(),
-        })),
-    ))
-}
-
-/// Get specific issue details
-pub async fn issue_detail(
-    State(storage): State<Arc<StorageManager>>,
-    Path(id): Path<Uuid>,
-) -> Result<Json<Value>> {
-    let issue = storage
-        .issues()
-        .find_by_id(id)
-        .await?
-        .ok_or_else(|| Error::NotFound(format!("Issue with id {}", id)))?;
-
-    Ok(Json(json!({
-        "issue": issue,
-        "timestamp": chrono::Utc::now(),
-    })))
-}
-
-/// Update an existing issue
-pub async fn issue_update(
-    State(storage): State<Arc<StorageManager>>,
-    State(ws_manager): State<Arc<WebSocketManager>>,
-    Path(id): Path<Uuid>,
-    Json(request): Json<IssueRequest>,
-) -> Result<Json<Value>> {
-    let mut issue = storage
-        .issues()
-        .find_by_id(id)
-        .await?
-        .ok_or_else(|| Error::NotFound(format!("Issue with id {}", id)))?;
-
-    // Update issue fields
-    issue.title = request.title;
-    issue.description = request.description;
-    issue.priority = match request.priority.as_str() {
-        "Low" => IssuePriority::Low,
-        "Medium" => IssuePriority::Medium,
-        "High" => IssuePriority::High,
-        "Critical" => IssuePriority::Critical,
-        _ => return Err(Error::BadRequest("Invalid priority".to_string())),
-    };
-    issue.assigned_agent_id = request.assigned_agent_id;
-    issue.updated_at = chrono::Utc::now();
-
-    storage.issues().update(&issue).await?;
-
-    // Broadcast to WebSocket clients
-    let _ = ws_manager.broadcast_issue_status(issue.id, format!("{:?}", issue.status));
-
-    Ok(Json(json!({
-        "issue": issue,
-        "timestamp": chrono::Utc::now(),
-    })))
-}
-
-/// Delete an issue
-pub async fn issue_delete(
-    State(storage): State<Arc<StorageManager>>,
-    Path(id): Path<Uuid>,
-) -> Result<StatusCode> {
-    let issue = storage
-        .issues()
-        .find_by_id(id)
-        .await?
-        .ok_or_else(|| Error::NotFound(format!("Issue with id {}", id)))?;
-
-    storage.issues().delete(id).await?;
-
-    Ok(StatusCode::NO_CONTENT)
-}
-
-// ======================
-// KNOWLEDGE API ENDPOINTS
-// ======================
 
 /// Knowledge entry creation request
 #[derive(Debug, Deserialize)]
