+++ vibe-ensemble-web/src/server.rs
@@ -2,7 +2,7 @@
 
 use crate::{handlers, Result};
 use axum::{
-    routing::{get, post},
+    routing::{delete, get, post, put},
     Router,
 };
 use std::sync::Arc;
@@ -46,17 +46,20 @@ impl WebServer {
             // Dashboard routes
             .route("/", get(handlers::dashboard))
             .route("/dashboard", get(handlers::dashboard))
-            
-            // API routes  
+            // API routes
             .route("/api/health", get(handlers::health))
+            .route("/api/stats", get(handlers::system_stats))
+            // Agent API routes
             .route("/api/agents", get(handlers::agents_list))
+            .route("/api/agents/:id", get(handlers::agent_get))
+            // Issue API routes
             .route("/api/issues", get(handlers::issues_list))
             .route("/api/issues", post(handlers::issues_create))
-            .route("/api/stats", get(handlers::system_stats))
-            
+            .route("/api/issues/:id", get(handlers::issue_get))
+            .route("/api/issues/:id", put(handlers::issue_update))
+            .route("/api/issues/:id", delete(handlers::issue_delete))
             // Add shared state
             .with_state(self.storage.clone())
-            
             // Add middleware
             .layer(
                 ServiceBuilder::new()
