+++ vibe-ensemble-web/src/templates.rs
